apps/api/requirements.txt
FastAPI
uvicorn
psycopg2-binary
pydantic[email]
sqlalchemy
alembic
httpx
python-dotenv
pytest
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-multipart